+++ src/index.ts
@@ -1,0 +1,3 @@
+import UbirchVerification from './verification';
+
+export { UbirchVerification };
