+++ src/verification/verification.ts
@@ -1,10 +1,10 @@
 import i18n from 'i18next';
 import { sha256 } from 'js-sha256';
 import { sha512 } from 'js-sha512';
-import * as BlockchainSettings from './blockchain-assets/blockchain-settings.json';
-import * as de from './assets/i18n/de.json';
-import * as en from './assets/i18n/en.json';
-import environment from './environment';
+import * as BlockchainSettings from '../blockchain-assets/blockchain-settings.json';
+import * as de from '../assets/i18n/de.json';
+import * as en from '../assets/i18n/en.json';
+import environment from '../environment';
 import { BehaviorSubject, Observable } from 'rxjs';
 import {
   EError,
@@ -24,7 +24,7 @@ import {
   IUbirchVerificationConfig,
   IUbirchVerificationResponse,
   IUbirchVerificationResult,
-} from './models';
+} from '../models';
 
 i18n.init(
   {
